requirements.txt
# Core dependencies for script functionality
requests>=2.31.0
aiohttp>=3.8.0
asyncio-throttle>=1.0.2

# API clients for multi-provider support
openai>=1.0.0
anthropic>=0.7.0
google-generativeai>=0.3.0

# Configuration and environment management
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Logging and monitoring
structlog>=23.0.0
rich>=13.0.0

# File and path utilities
pathlib2>=2.3.7
watchdog>=3.0.0

# Web framework for backend
fastapi>=0.104.0
uvicorn>=0.24.0

# Database and storage
motor>=3.3.0
beanie>=1.23.0
redis>=5.0.0

# Additional utilities
click>=8.1.0
colorama>=0.4.6
tabulate>=0.9.0

# Testing (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Development tools (optional)
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0